Added a dependency on 'try-catch' as part of moving away from handy
The `handy` module (which is mine) is too cluttered and needs to be
broken out.  I split handy/try off into a separate try-catch module
and am migrating things to using that instead.  Eventually
struct-plus-plus will use a variety of dedicated modules instead of
the bloated Fibber McGee that is `handy`.